Print bone mciroCT comparison analysis results

This function is a wraps and extends knitr::kable() to format and print the results of bone microCT comparison analyses at one site by sex for use in an R Markdown document. A code chunk containing this function should probably have the results = "asis" option to allow pretty formatting. Any NA values are printed as an empty string.

Usage

print_results(results, ...)

Arguments

results

A list of microCT comparison results, formatted as is the output of compare_genotypes().

...

Additional arguments passed on to knitr::kable(). This function uses a local default of digits = 3 to specify how many significant digits to print. This can be modified by passing a user specified digits value.

Value

Text output which is by default a Pandoc markdown pipe table for each measure and each sex analyzed. If two sexes are analyzed, Pandoc fenced div syntax is used to print the tables in two columns by sex. For example:

:::: {style="display: flex;"}

::: {}

**BV/TV**

|Sex |Genotype |  n| Mean|  SEM|     P|Sig |
|:---|:--------|--:|----:|----:|-----:|:---|
|M   |Cre      |  3| 18.5| 1.58|      |    |
|M   |Cre;fl   |  3| 19.9| 1.11| 0.701|    |

:::

::: {}

**BV/TV**

|Sex |Genotype |  n| Mean|   SEM|     P|Sig |
|:---|:--------|--:|----:|-----:|-----:|:---|
|F   |Cre      |  3| 18.3| 1.041|      |    |
|F   |Cre;fl   |  3| 17.1| 0.936| 0.656|    |

:::

::::

Examples

gen_key <- read_key_csv(mctr_ex("example-gen-key.csv"))
#> Rows: 12 Columns: 7
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (2): Sex, Genotype
#> dbl (5): AS, SampNo, Spine, Met, Dia
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
gen_trab <- read_trabecular_csv(mctr_ex("example-trabecular.csv"),
                            gen_key)
#> Rows: 24 Columns: 75
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (19): SampName, MeasDate, ListDate, Filename, Site, Energy-I-Code, Contr...
#> dbl (51): SampNo, MeasNo, Integr.Time, Sigma, Support, Threshold, Unit, Data...
#> lgl  (5): S-DOB, S-Remark, Meas-Rmk, RAW-Label, IMA-Label
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
Sp.trab <- gen_trab |> dplyr::filter(Site == "Spine") |> compare_genotypes()
print_results(Sp.trab)
#> :::: {style="display: flex;"}
#> 
#> ::: {}
#> 
#> **BV/TV**
#> 
#> |Sex |Genotype |  n| Mean|  SEM|     P|Sig |
#> |:---|:--------|--:|----:|----:|-----:|:---|
#> |M   |Cre      |  3| 18.5| 2.73|      |    |
#> |M   |Cre;fl   |  3| 19.9| 1.92| 0.701|    |
#> 
#> 
#> **SMI**
#> 
#> |Sex |Genotype |  n|  Mean|   SEM|     P|Sig |
#> |:---|:--------|--:|-----:|-----:|-----:|:---|
#> |M   |Cre      |  3| 1.209| 0.215|      |    |
#> |M   |Cre;fl   |  3| 0.796| 0.128| 0.174|    |
#> 
#> 
#> **Tb.N**
#> 
#> |Sex |Genotype |  n| Mean|   SEM|     P|Sig |
#> |:---|:--------|--:|----:|-----:|-----:|:---|
#> |M   |Cre      |  3| 3.36| 0.135|      |    |
#> |M   |Cre;fl   |  3| 3.39| 0.334| 0.938|    |
#> 
#> 
#> **Tb.Th**
#> 
#> |Sex |Genotype |  n|  Mean|   SEM|     P|Sig |
#> |:---|:--------|--:|-----:|-----:|-----:|:---|
#> |M   |Cre      |  3| 0.061| 0.005|      |    |
#> |M   |Cre;fl   |  3| 0.059| 0.001| 0.782|    |
#> 
#> 
#> **Tb.Sp**
#> 
#> |Sex |Genotype |  n|  Mean|   SEM|     P|Sig |
#> |:---|:--------|--:|-----:|-----:|-----:|:---|
#> |M   |Cre      |  3| 0.297| 0.009|      |    |
#> |M   |Cre;fl   |  3| 0.299| 0.032| 0.963|    |
#> 
#> 
#> :::
#> 
#> ::: {}
#> 
#> **BV/TV**
#> 
#> |Sex |Genotype |  n| Mean|  SEM|     P|Sig |
#> |:---|:--------|--:|----:|----:|-----:|:---|
#> |F   |Cre      |  3| 18.3| 1.80|      |    |
#> |F   |Cre;fl   |  3| 17.1| 1.62| 0.656|    |
#> 
#> 
#> **SMI**
#> 
#> |Sex |Genotype |  n|  Mean|   SEM|     P|Sig |
#> |:---|:--------|--:|-----:|-----:|-----:|:---|
#> |F   |Cre      |  3| 0.925| 0.083|      |    |
#> |F   |Cre;fl   |  3| 0.852| 0.157| 0.704|    |
#> 
#> 
#> **Tb.N**
#> 
#> |Sex |Genotype |  n| Mean|   SEM|     P|Sig |
#> |:---|:--------|--:|----:|-----:|-----:|:---|
#> |F   |Cre      |  3| 6.65| 3.802|      |    |
#> |F   |Cre;fl   |  3| 3.05| 0.248| 0.443|    |
#> 
#> 
#> **Tb.Th**
#> 
#> |Sex |Genotype |  n|  Mean|   SEM|    P|Sig |
#> |:---|:--------|--:|-----:|-----:|----:|:---|
#> |F   |Cre      |  3| 0.053| 0.004|     |    |
#> |F   |Cre;fl   |  3| 0.058| 0.001| 0.34|    |
#> 
#> 
#> **Tb.Sp**
#> 
#> |Sex |Genotype |  n|  Mean|   SEM|    P|Sig |
#> |:---|:--------|--:|-----:|-----:|----:|:---|
#> |F   |Cre      |  3| 0.287| 0.098|     |    |
#> |F   |Cre;fl   |  3| 0.335| 0.031| 0.66|    |
#> 
#> 
#> :::
#> 
#> ::::
#>