Read in finite element analysis data

This is a wrapper around readr::read_csv() to read in and process finite element analysis microCT data from a mfe.csv file. It uses a key object to look up and add sex, genotype, and site information.

Usage

read_mfe_csv(file, key, ...)

Arguments

file

The file path to the mfe.csv file.

key

The key object containing sample information, as created by the read_key_csv() function.

...

Additional arguments passed on to readr::read_csv().

Value

A tibble/data frame containing finite element analysis data. For example:

|   AS|Sex |Genotype | SampNo|Site  | MeasNo|       S|  F.ult|
|----:|:---|:--------|------:|:-----|------:|-------:|------:|
| 1365|M   |Cre      |  10778|Spine |  31711| 11480.0| 32.257|
| 1365|M   |Cre      |  10778|Met   |  31710| 19034.1| 57.483|
| 1365|M   |Cre      |  10778|Dia   |  31712| 25405.7| 79.100|
| 1366|F   |Cre      |  10779|Spine |  31714| 12449.0| 34.213|
| 1366|F   |Cre      |  10779|Met   |  31713| 20679.9| 62.022|
| 1366|F   |Cre      |  10779|Dia   |  31715| 22832.0| 70.721|

Examples

key <- read_key_csv(microCTr_example("example-key.csv"))
#> Rows: 12 Columns: 7
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (2): Sex, Genotype
#> dbl (5): AS, SampNo, Spine, Met, Dia
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
mfe <- read_mfe_csv(microCTr_example("example-mfe.csv"),
                                     key)
#> Rows: 36 Columns: 47
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (12): SampName, MeasDate, ListDate, Filename, Site, Energy-I-Code, Contr...
#> dbl (30): SampNo, MeasNo, Integr.Time, Sigma, Support, Threshold, Unit, Data...
#> lgl  (5): S-DOB, S-Remark, Meas-Rmk, RAW-Label, IMA-Label
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.